Completely moved the character page to its own branch of the URL, deleted all old files, and added a create endpoint
