+++ src/templates/character/character.html
@@ -3,13 +3,42 @@
 {% block title_content %}{{ character_id }}'s Profile{% endblock %}
 
 {% block body_content %}
+    <div class="main-content">
+        {% if current_user.is_authenticated %}
+        <h1>{{ character.name }}</h1>
+        <p>Owner: {{ user.username }}</p>
 
+        <form method="POST" action="{{ url_for('character_bp.character', character_id=character.char_id) }}">
+            <label for="class_id">Choose a Class:</label>
+            <select id="class_id" name="class_id">
+                {% for dnd_class in all_classes %}
+                    <option value="{{ dnd_class.class_id }}"
+                        {% if current_class and dnd_class.name == current_class %} selected {% endif %}>
+                        {{ dnd_class.name }}
+                    </option>
+                {% endfor %}
+            </select>
 
-{% if current_user.is_authenticated %}
+            <label for="class_level">Set Class Level:</label>
+            <input type="number" id="class_level" name="class_level" min="1" value="{{ current_level if current_level else 1 }}">
 
-<h1>character page: {{ character_id }}</h1>
-{% else %}
-<h1>current user not found</h1>
+            <button type="submit">Update</button>
+        </form>
 
-{%endif%}
+        <p>
+            Current Class: {{ current_class if current_class else "None" }} <br>
+            Current Level: {{ current_level if current_level else "N/A" }}
+        </p>
+
+            <a href="/profile/{{current_user.username}}">Back to Profile</a>
+
+        {% else %}
+        <h1>You're not logged in!</h1>
+        <a href="{{ url_for('auth_bp.login') }}">Click here to log in.<br></a>
+        <a href="{{ url_for('auth_bp.register') }}">Or, you can click here to create a new account!</a>
+
+        {%endif%}
+    </div>
+
+    
 {% endblock %}
